requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
fastremap==1.9.2
fill-voids==1.1.0
gast==0.3.3
grpcio==1.27.2
h5py==2.10.0
httplib2==0.9.2
idna==2.6
imageio==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.31.0
louis==3.5.0
lungmask==0.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==3.1.3
mock==4.0.1
netifaces==0.10.4
networkx==2.4
nibabel==3.0.1
numba==0.48.0
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
opencv-python==4.2.0.32
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.11.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydicom==1.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.16.2
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
SimpleITK==1.2.4
simplejson==3.13.2
six==1.14.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
torch==1.4.0
tqdm==4.43.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.0
xkit==0.0.0
zope.interface==4.3.2
